Conflicts:
	app/controllers/lp/posts_controller.rb
	config/routes.rb
	lib/auth/lp_current_user_provider.rb

+++ app/controllers/posts_controller.rb
@@ -5,7 +5,7 @@ require_dependency 'distributed_memoizer'
 class PostsController < ApplicationController
 
   # Need to be logged in for all actions here
-  before_filter :ensure_logged_in, except: [:show, :replies, :by_number, :short_link, :versions, :reply_history]
+  before_filter :ensure_logged_in, except: [:show, :replies, :by_number, :short_link, :reply_history, :revisions, :expand_embed]
 
   skip_before_filter :store_incoming_links, only: [:short_link]
   skip_before_filter :check_xhr, only: [:markdown,:short_link]
@@ -29,8 +29,9 @@ class PostsController < ApplicationController
     params = create_params
 
     key = params_key(params)
+    error_json = nil
 
-    result = DistributedMemoizer.memoize(key, 120) do
+    payload = DistributedMemoizer.memoize(key, 120) do
       post_creator = PostCreator.new(current_user, params)
       post = post_creator.create
       if post_creator.errors.present?
@@ -38,37 +39,43 @@ class PostsController < ApplicationController
         # If the post was spam, flag all the user's posts as spam
         current_user.flag_linked_posts_as_spam if post_creator.spam?
 
-        "e" << MultiJson.dump(errors: post_creator.errors.full_messages)
+        error_json = MultiJson.dump(errors: post_creator.errors.full_messages)
+        raise Discourse::InvalidPost
+
       else
         post_serializer = PostSerializer.new(post, scope: guardian, root: false)
         post_serializer.topic_slug = post.topic.slug if post.topic.present?
         post_serializer.draft_sequence = DraftSequence.current(current_user, post.topic.draft_key)
-        "s" << MultiJson.dump(post_serializer)
+        MultiJson.dump(post_serializer)
       end
     end
 
+    render json: payload
 
-    payload = result[1..-1]
-    if result[0] == "e"
-      # don't memoize errors
-      $redis.del(DistributedMemoizer.redis_key(key))
-      render json: payload, status: 422
-    else
-      render json: payload
-    end
+  rescue Discourse::InvalidPost
+    render json: error_json, status: 422
   end
 
   def update
     params.require(:post)
 
-    post = Post.where(id: params[:id]).first
+    post = Post.where(id: params[:id])
+    post = post.with_deleted if guardian.is_staff?
+    post = post.first
     post.image_sizes = params[:image_sizes] if params[:image_sizes].present?
+
+    if too_late_to(:edit, post)
+      render json: {errors: [I18n.t('too_late_to_edit')]}, status: 422
+      return
+    end
+
     guardian.ensure_can_edit!(post)
 
     # to stay consistent with the create api,
     #  we should allow for title changes and category changes here
-    # we should also move all of this to a post updater.
+    #  we should also move all of this to a post updater.
     if post.post_number == 1 && (params[:title] || params[:post][:category])
+      post.topic.acting_user = current_user
       post.topic.title = params[:title] if params[:title]
       Topic.transaction do
         post.topic.change_category(params[:post][:category])
@@ -82,10 +89,9 @@ class PostsController < ApplicationController
     end
 
     revisor = PostRevisor.new(post)
-    if revisor.revise!(current_user, params[:post][:raw])
+    if revisor.revise!(current_user, params[:post][:raw], edit_reason: params[:post][:edit_reason])
       TopicLink.extract_from(post)
     end
-
 
     if post.errors.present?
       render_json_error(post)
@@ -106,34 +112,41 @@ class PostsController < ApplicationController
     render_json_dump(result)
   end
 
+  def show
+    post = find_post_from_params
+    display_post(post)
+  end
+
   def by_number
-    @post = Post.where(topic_id: params[:topic_id], post_number: params[:post_number]).first
-    guardian.ensure_can_see!(@post)
-    @post.revert_to(params[:version].to_i) if params[:version].present?
-    render_post_json(@post)
+    post = find_post_from_params_by_number
+    display_post(post)
   end
 
   def reply_history
-    @post = Post.where(id: params[:id]).first
-    guardian.ensure_can_see!(@post)
-
-    render_serialized(@post.reply_history, PostSerializer)
-  end
-
-  def show
-    @post = find_post_from_params
-    @post.revert_to(params[:version].to_i) if params[:version].present?
-    render_post_json(@post)
+    post = find_post_from_params
+    render_serialized(post.reply_history, PostSerializer)
   end
 
   def destroy
     post = find_post_from_params
+
+    if too_late_to(:delete_post, post)
+      render json: {errors: [I18n.t('too_late_to_edit')]}, status: 422
+      return
+    end
+
     guardian.ensure_can_delete!(post)
 
     destroyer = PostDestroyer.new(current_user, post)
     destroyer.destroy
 
     render nothing: true
+  end
+
+  def expand_embed
+    render json: {cooked: TopicEmbed.expanded_for(find_post_from_params) }
+  rescue
+    render_json_error I18n.t('errors.embed.load_from_remote')
   end
 
   def recover
@@ -157,23 +170,23 @@ class PostsController < ApplicationController
     posts.each {|p| guardian.ensure_can_delete!(p) }
 
     Post.transaction do
-      topic_id = posts.first.topic_id
       posts.each {|p| PostDestroyer.new(current_user, p).destroy }
     end
 
     render nothing: true
   end
 
-  # Retrieves a list of versions and who made them for a post
-  def versions
-    post = find_post_from_params
-    render_serialized(post.all_versions, VersionSerializer)
-  end
-
   # Direct replies to this post
   def replies
     post = find_post_from_params
     render_serialized(post.replies, PostSerializer)
+  end
+
+  def revisions
+    post_revision = find_post_revision_from_params
+    guardian.ensure_can_see!(post_revision)
+    post_revision_serializer = PostRevisionSerializer.new(post_revision, scope: guardian, root: false)
+    render_json_dump(post_revision_serializer)
   end
 
   def bookmark
@@ -188,64 +201,99 @@ class PostsController < ApplicationController
     render nothing: true
   end
 
-
   protected
 
-    def find_post_from_params
-      finder = Post.where(id: params[:id] || params[:post_id])
+  def find_post_revision_from_params
+    post_id = params[:id] || params[:post_id]
+    revision = params[:revision].to_i
+    raise Discourse::InvalidParameters.new(:revision) if revision < 2
 
-      # Include deleted posts if the user is staff
-      finder = finder.with_deleted if current_user.try(:staff?)
+    post_revision = PostRevision.where(post_id: post_id, number: revision).first
+    post_revision.post = find_post_from_params
 
-      post = finder.first
-      guardian.ensure_can_see!(post)
-      post
-    end
+    guardian.ensure_can_see!(post_revision)
+    post_revision
+  end
 
   def render_post_json(post)
     post_serializer = PostSerializer.new(post, scope: guardian, root: false)
     post_serializer.add_raw = true
+    counts = PostAction.counts_for([post], current_user)
+    if counts && counts = counts[post.id]
+      post_serializer.post_actions = counts
+    end
     render_json_dump(post_serializer)
   end
 
   private
 
-    def params_key(params)
-      "post##" << Digest::SHA1.hexdigest(params
-        .to_a
-        .concat([["user", current_user.id]])
-        .sort{|x,y| x[0] <=> y[0]}.join do |x,y|
-          "#{x}:#{y}"
-        end)
+  def params_key(params)
+    "post##" << Digest::SHA1.hexdigest(params
+      .to_a
+      .concat([["user", current_user.id]])
+      .sort{|x,y| x[0] <=> y[0]}.join do |x,y|
+        "#{x}:#{y}"
+      end)
+  end
+
+  def create_params
+    permitted = [
+      :raw,
+      :topic_id,
+      :title,
+      :archetype,
+      :category,
+      :target_usernames,
+      :reply_to_post_number,
+      :auto_close_time,
+      :auto_track
+    ]
+
+    # param munging for WordPress
+    params[:auto_track] = !(params[:auto_track].to_s == "false") if params[:auto_track]
+
+    if api_key_valid?
+      # php seems to be sending this incorrectly, don't fight with it
+      params[:skip_validations] = params[:skip_validations].to_s == "true"
+      permitted << :skip_validations
+
+      # We allow `embed_url` via the API
+      permitted << :embed_url
     end
 
-    def create_params
-      permitted = [
-        :raw,
-        :topic_id,
-        :title,
-        :archetype,
-        :category,
-        :target_usernames,
-        :reply_to_post_number,
-        :auto_close_days,
-        :auto_track
-      ]
-
-      # param munging for WordPress
-      params[:auto_track] = !(params[:auto_track].to_s == "false") if params[:auto_track]
-
-      if api_key_valid?
-        # php seems to be sending this incorrectly, don't fight with it
-        params[:skip_validations] = params[:skip_validations].to_s == "true"
-        permitted << :skip_validations
-      end
-
-      params.require(:raw)
-      params.permit(*permitted).tap do |whitelisted|
-          whitelisted[:image_sizes] = params[:image_sizes]
-          # TODO this does not feel right, we should name what meta_data is allowed
-          whitelisted[:meta_data] = params[:meta_data]
-      end
+    params.require(:raw)
+    params.permit(*permitted).tap do |whitelisted|
+        whitelisted[:image_sizes] = params[:image_sizes]
+        # TODO this does not feel right, we should name what meta_data is allowed
+        whitelisted[:meta_data] = params[:meta_data]
     end
+  end
+
+  def too_late_to(action, post)
+    !guardian.send("can_#{action}?", post) && post.user_id == current_user.id && post.edit_time_limit_expired?
+  end
+
+  def display_post(post)
+    post.revert_to(params[:version].to_i) if params[:version].present?
+    render_post_json(post)
+  end
+
+  def find_post_from_params
+    by_id_finder = Post.where(id: params[:id] || params[:post_id])
+    find_post_using(by_id_finder)
+  end
+
+  def find_post_from_params_by_number
+    by_number_finder = Post.where(topic_id: params[:topic_id], post_number: params[:post_number])
+    find_post_using(by_number_finder)
+  end
+
+  def find_post_using(finder)
+    # Include deleted posts if the user is staff
+    finder = finder.with_deleted if current_user.try(:staff?)
+    post = finder.first
+    guardian.ensure_can_see!(post)
+    post
+  end
+
 end
