+++ app/models/category_featured_topic.rb
@@ -5,18 +5,25 @@ class CategoryFeaturedTopic < ActiveRecord::Base
   # Populates the category featured topics
   def self.feature_topics
     transaction do
-      Category.all.each do |c|
-        feature_topics_for(c)
-        CategoryFeaturedUser.feature_users_in(c)
+      current = {}
+      CategoryFeaturedTopic.select(:topic_id, :category_id).order(:rank).each do |f|
+        (current[f.category_id] ||= []) << f.topic_id
+      end
+      Category.select(:id, :topic_id).all.each do |c|
+
+        feature_topics_for(c, current[c.id] || [])
+        CategoryFeaturedUser.feature_users_in(c.id)
       end
     end
   end
 
-  def self.feature_topics_for(c)
+  def self.feature_topics_for(c, existing=nil)
     return if c.blank?
 
-    query = TopicQuery.new(self.fake_admin, per_page: SiteSetting.category_featured_topics, except_topic_id: c.topic_id, visible: true)
-    results = query.list_category(c).topic_ids.to_a
+    query = TopicQuery.new(self.fake_admin, per_page: SiteSetting.category_featured_topics, except_topic_ids: [c.topic_id], visible: true, no_definitions: true)
+    results = query.list_category(c).topic_ids
+
+    return if results == existing
 
     CategoryFeaturedTopic.transaction do
       CategoryFeaturedTopic.delete_all(category_id: c.id)
@@ -45,8 +52,8 @@ end
 #
 #  category_id :integer          not null
 #  topic_id    :integer          not null
-#  created_at  :datetime         not null
-#  updated_at  :datetime         not null
+#  created_at  :datetime
+#  updated_at  :datetime
 #  rank        :integer          default(0), not null
 #  id          :integer          not null, primary key
 #
@@ -55,4 +62,3 @@ end
 #  cat_featured_threads                                    (category_id,topic_id) UNIQUE
 #  index_category_featured_topics_on_category_id_and_rank  (category_id,rank)
 #
-
